Move agent access UI into client agent
Keep the access manifest independently composable without adding a second
published package or Storybook deployment.

Refs ENG-1842
