+++ .changeset/agent-access-client-ui.md
@@ -1,5 +1,5 @@
 ---
-'@shipfox/client-agent-access': minor
+'@shipfox/client-agent': minor
 ---
 
 Add the Agent access consent and workspace credential-management UI.
