+++ .github/workflows/ci.yml
@@ -326,7 +326,6 @@ jobs:
             --filter=@shipfox/api... \
             --filter=@shipfox/client-shell... \
             --filter=@shipfox/client-features... \
-            --filter=@shipfox/client-agent-access... \
             --concurrency="$SHIPFOX_TURBO_CONCURRENCY"
 
       - name: Verify packed API runtime closure
