+++ apps/storybook/preview-manifest.ts
@@ -21,19 +21,10 @@ export const storybooks = [
     url: '/react-ui/index.html',
   },
   {
-    id: 'client-agent-access',
-    title: 'Client Agent Access',
-    package: '@shipfox/client-agent-access',
-    order: 2,
-    source: 'libs/client/agent-access/storybook-static',
-    path: '/client-agent-access/',
-    url: '/client-agent-access/index.html',
-  },
-  {
     id: 'client-agent',
     title: 'Client Agent',
     package: '@shipfox/client-agent',
-    order: 3,
+    order: 2,
     source: 'libs/client/agent/storybook-static',
     path: '/client-agent/',
     url: '/client-agent/index.html',
@@ -42,7 +33,7 @@ export const storybooks = [
     id: 'client-auth',
     title: 'Client Auth',
     package: '@shipfox/client-auth',
-    order: 4,
+    order: 3,
     source: 'libs/client/auth/storybook-static',
     path: '/client-auth/',
     url: '/client-auth/index.html',
@@ -51,7 +42,7 @@ export const storybooks = [
     id: 'client-integrations',
     title: 'Client Integrations',
     package: '@shipfox/client-integrations',
-    order: 5,
+    order: 4,
     source: 'libs/client/integrations/storybook-static',
     path: '/client-integrations/',
     url: '/client-integrations/index.html',
@@ -60,7 +51,7 @@ export const storybooks = [
     id: 'client-logs',
     title: 'Client Logs',
     package: '@shipfox/client-logs',
-    order: 6,
+    order: 5,
     source: 'libs/client/logs/storybook-static',
     path: '/client-logs/',
     url: '/client-logs/index.html',
@@ -69,7 +60,7 @@ export const storybooks = [
     id: 'client-onboarding',
     title: 'Client Onboarding',
     package: '@shipfox/client-onboarding',
-    order: 7,
+    order: 6,
     source: 'libs/client/onboarding/storybook-static',
     path: '/client-onboarding/',
     url: '/client-onboarding/index.html',
@@ -78,7 +69,7 @@ export const storybooks = [
     id: 'client-projects',
     title: 'Client Projects',
     package: '@shipfox/client-projects',
-    order: 8,
+    order: 7,
     source: 'libs/client/projects/storybook-static',
     path: '/client-projects/',
     url: '/client-projects/index.html',
@@ -87,7 +78,7 @@ export const storybooks = [
     id: 'client-runners',
     title: 'Client Runners',
     package: '@shipfox/client-runners',
-    order: 9,
+    order: 8,
     source: 'libs/client/runners/storybook-static',
     path: '/client-runners/',
     url: '/client-runners/index.html',
@@ -96,7 +87,7 @@ export const storybooks = [
     id: 'client-secrets',
     title: 'Client Secrets',
     package: '@shipfox/client-secrets',
-    order: 10,
+    order: 9,
     source: 'libs/client/secrets/storybook-static',
     path: '/client-secrets/',
     url: '/client-secrets/index.html',
@@ -105,7 +96,7 @@ export const storybooks = [
     id: 'client-triggers',
     title: 'Client Triggers',
     package: '@shipfox/client-triggers',
-    order: 11,
+    order: 10,
     source: 'libs/client/triggers/storybook-static',
     path: '/client-triggers/',
     url: '/client-triggers/index.html',
@@ -114,7 +105,7 @@ export const storybooks = [
     id: 'client-workflows',
     title: 'Client Workflows',
     package: '@shipfox/client-workflows',
-    order: 12,
+    order: 11,
     source: 'libs/client/workflows/storybook-static',
     path: '/client-workflows/',
     url: '/client-workflows/index.html',
