+++ cloudflare-pages.config.json
@@ -17,7 +17,6 @@
       "target": "@shipfox/storybook",
       "affectedTargets": [
         "@shipfox/react-ui",
-        "@shipfox/client-agent-access",
         "@shipfox/client-agent",
         "@shipfox/client-auth",
         "@shipfox/client-integrations",
@@ -34,11 +33,6 @@
         "metadataPath": "/preview-metadata.json",
         "endpoints": [
           { "id": "react-ui", "path": "/react-ui/index.json", "requireNonEmpty": true },
-          {
-            "id": "client-agent-access",
-            "path": "/client-agent-access/index.json",
-            "requireNonEmpty": true
-          },
           { "id": "client-agent", "path": "/client-agent/index.json", "requireNonEmpty": true },
           { "id": "client-auth", "path": "/client-auth/index.json", "requireNonEmpty": true },
           {
