--- libs/client/agent-access/.storybook/main.ts
@@ -1,1 +1,0 @@
-export {default} from '../../.storybook/main.ts';
