--- libs/client/agent-access/.storybook/preview.css
@@ -1,5 +1,0 @@
-@import "@shipfox/react-ui/index.css";
-
-@source "../src";
-@source "../../shell/src";
-@source "../../../shared/react/ui/src";
