--- libs/client/agent-access/CHANGELOG.md
@@ -1,1 +1,0 @@
-# @shipfox/client-agent-access
