--- libs/client/agent-access/test/setup.ts
@@ -1,3 +1,0 @@
-import {installClientDomTestEnv} from '@shipfox/client-test-setup';
-
-installClientDomTestEnv();
