--- libs/client/agent-access/tsconfig.build.json
@@ -1,9 +1,0 @@
-{
-  "extends": "@shipfox/ts-config/react",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist"
-  },
-  "include": ["src"],
-  "exclude": ["**/*.test.tsx", "**/*.test.ts"]
-}
