--- libs/client/agent-access/vercel.json
@@ -1,11 +1,0 @@
-{
-  "$schema": "https://openapi.vercel.sh/vercel.json",
-  "buildCommand": "turbo build && pnpm run storybook:build",
-  "devCommand": "pnpm run storybook",
-  "installCommand": "pnpm install",
-  "git": {
-    "deploymentEnabled": false
-  },
-  "framework": null,
-  "outputDirectory": "./storybook-static"
-}
