+++ libs/client/agent/.storybook/preview.css
@@ -1,4 +1,5 @@
 @import "@shipfox/react-ui/index.css";
 
 @source "../src";
+@source "../../shell/src";
 @source "../../../shared/react/ui/src";
