+++ libs/client/agent/package.json
@@ -63,6 +63,7 @@
   },
   "dependencies": {
     "@shipfox/api-agent-dto": "workspace:*",
+    "@shipfox/api-auth-dto": "workspace:*",
     "@shipfox/client-api": "workspace:*",
     "@shipfox/client-shell": "workspace:*",
     "@shipfox/client-ui": "workspace:*",
