+++ libs/client/agent/src/agent-access/components/format.ts
