+++ libs/client/agent/src/agent-access/components/oauth-consent-page.tsx
@@ -8,13 +8,13 @@ import {RadioGroup, RadioGroupItem} from '@shipfox/react-ui/radio-group';
 import {Skeleton} from '@shipfox/react-ui/skeleton';
 import {Code, Text} from '@shipfox/react-ui/typography';
 import {useEffect, useState} from 'react';
-import type {OAuthConsent} from '#core/agent-access.js';
+import type {OAuthConsent} from '#agent-access/core/agent-access.js';
 import {
   useApproveOAuthConsentMutation,
   useDenyOAuthConsentMutation,
   useOAuthConsentQuery,
-} from '#hooks/api/consent.js';
-import {validateOAuthConsentSearch} from '#routes/inputs.js';
+} from '#agent-access/hooks/api/consent.js';
+import {validateOAuthConsentSearch} from '#agent-access/routes/inputs.js';
 import {oauthConsentErrorMessage} from './errors.js';
 import {formatAgentAccessTimestamp} from './format.js';
