+++ libs/client/agent/src/routes/agent-access-settings.tsx
@@ -1,6 +1,6 @@
 import {defineRoute, useActiveWorkspace} from '@shipfox/client-shell/runtime';
 import {Header} from '@shipfox/react-ui/typography';
-import {AgentAccessSettingsPage} from '#components/agent-access-settings-page.js';
+import {AgentAccessSettingsPage} from '#agent-access/components/agent-access-settings-page.js';
 
 export default defineRoute({
   staticData: {frame: 'content'},
