+++ publication-closure.json
@@ -16,7 +16,6 @@
     "@shipfox/api-triggers",
     "@shipfox/api-workflows",
     "@shipfox/api-workspaces",
-    "@shipfox/client-agent-access",
     "@shipfox/client-agent",
     "@shipfox/client-auth",
     "@shipfox/client-features",
@@ -81,7 +80,6 @@
     "@shipfox/api-workflows-dto",
     "@shipfox/api-workspaces",
     "@shipfox/api-workspaces-dto",
-    "@shipfox/client-agent-access",
     "@shipfox/client-agent",
     "@shipfox/client-api",
     "@shipfox/client-auth",
